simpler torch, test on ubuntu

+++ experience_replay.py
@@ -24,9 +24,8 @@ class ExperienceReplay:
   def sample_batch(self, batch_size):
     """
     Sample random experience according to batch size from memory
-
     Returns a np array of each
     """
     indices = np.random.choice(len(self.memory), batch_size, replace=False)
     states, actions, rewards, next_states, dones = zip(*[self.memory[i] for i in indices])
-    return np.array(states), np.array(actions), np.array(rewards, dtype=np.float32), np.array(next_states), np.array(dones, dtype=np.uint8)
+    return np.array(states), np.array(actions, dtype=np.int64), np.array(rewards, dtype=np.float32), np.array(next_states), np.array(dones, dtype=np.uint8)
